requirements.txt
streamlit
openai
pandas
chromadb
google-generativeai
python-dotenv
toml
pydantic
